// dev3-frontend-dev/services/api/dev3Schemas.ts
export type User = {
  /**
   * @format date-time
   */
  updatedAt: string;
  /**
   * @format date-time
   */
  createdAt: string;
  isCensored: boolean;
  isActive: boolean;
  uid: string;
  accountType: string;
  roles: string[];
  username: string;
  nearWalletAccountId: string;
};
// ...
export type ApiKey = {
  /**
   * @format date-time
   */
  updatedAt: string;
  /**
   * @format date-time
   */
  createdAt: string;
  api_key: string;
  /**
   * @format date-time
   */
  expires: string;
  owner: User;
};